solutions/34-farhan/TicTacToc/src/sample/ClassicTheme.java
package sample;

public class ClassicTheme {
    String currentTheme;
    boolean isClassicTheme = true;
    boolean isForrestTheme = false;
    boolean isHighContrastTheme = false;
    ClassicTheme(){
        currentTheme="classic";
    }

    public String getCurrentTheme() {
        return currentTheme;
    }

    public boolean isClassic(){
        return isClassicTheme;
    }

    public boolean isForrest(){
        return isForrestTheme;
    }

    public boolean isHighContrast(){
        return isHighContrastTheme;
    }

}
